chatbot/requirements.txt
aiogram
sqlalchemy
pydantic-settings
python-dotenv
